Remove a `session.commit()`
Reason: As a rule, when we get passed a session from a caller, it's the caller's job to decide whether the session will get rolled back or committed at the very end.

If we commit in a sub-method, the database state will look very weird to a caller that calls `rollback()` later on and expects everything to be rolled back.